TEXTURify/server/requirements.txt
pydantic==0.32.2
starlette==0.12.8
uvicorn==0.9.0
uvloop==0.13.0
fastapi==0.38.1
firebase==3.0.1
python-jwt==3.2.5
gcloud==0.18.3
sseclient==0.0.24
pycryptodome==3.4.3
requests-toolbelt==0.9.1
Pillow==6.2.1
numpy==1.17.4